fix decimal→int conversion
payment.amount (and refund.amount) are of type Decimal.

The correct way to convert to cents is to perform the multiplication with 100 in
the Decimal type and then convert to int:

  > int(Decimal('81.35') * 100)
  8135

Before this fix, the code inadvertantly stripped of one cent with certain
amounts, for example:

  > float(Decimal('81.35')) * 100
  8134.999999999999

Interestingly enough, the inverse operation float(cents) / 100 does seem to be
safe.